requirements.txt
aiohttp==3.8.3
python-telegram-bot==20.0a4
aioredis==2.0.1
hiredis==2.0.0
marshmallow==3.18.0
aiohttp-apispec==2.2.3
